fix(catalog-browser): convert watermark helper to TypeScript module
Previous attempt used dynamic import() of a .js file outside src/,
which TypeScript cannot resolve (TS2307: Cannot find module).

Solution: port the helper to TypeScript at src/services/watermark-helper.ts.
- import sharp from 'sharp' (already in package.json deps)
- Properly typed applyWatermark(imageBuffer: Buffer): Promise<Buffer>
- Compiled by tsc + bundled by tsup like other src/ files
- Import via relative path: ../../../../services/watermark-helper

Also: remove Docker/watermark/watermark-helper.js (replaced by TS version)
Add @services/* path alias to tsconfig.json for future use.

No functional change — watermark still applied to downloaded catalog
images with same specs (14pt, 40% opacity, bottom-right, EXIF metadata).
